Django-Web-App/requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
docopt==0.6.2
drf-writable-nested==0.6.0
Faker==4.0.3
idna==2.9
inflection==0.4.0
numpy==1.18.3
Pillow==7.1.1
pipreqs==0.4.10
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
yarg==0.1.9
